Move gi import at the beginning
This is for Windows. This ensure that the first DLL loaded is
libgirepository-1.0-1.dll. It has a fewer chance to exists in
c:\windows\system32 than zlib1.dll which was loaded by import shutil.
This is not perfect but it reduce the chance of conflicting DLLs. When
SafeDllSearchMode is enabled there is no way to prevent
c:\windows\system32 DLLs to be loaded first.

This is related to #300